clean up undo
there are a couple of important parts that work together: first, we delay the
keyboard event sent when the ICE commits a change. next, we track selection
events in the embedded sheet. and finally in the case of data changes, we cache
the tracked selection pending a recalculation.

it's a little convoluted but it all fits together and cleans up the UX of undo.

TODO: redo