update context to provide access to data model
to support finding the containing range of a formula, specifically for
HistogramTable. this used to work, because of an accidental side-effect,
which was subsequently fixed. so now we're explicit.

it is worth noting that giving functions access to the model could allow
functions to cause all sorts of havoc, so it might be worth proxying or
copying or something.

OTOH, we don't allow external functions, so maybe not an issue, or maybe
if we do allow external functions in the future we could lock down just
for those functions.