fix for issue with scale, charts, and tabs
if you viewed a tab with a chart, switched to another tab, then rescaled,
the scale would be off on the chart on the background tab. this is because
chart updates are as lazy as possible.

we now mark all charts as dirty on a rescale, so they can be updated the
next time they are added to layout.