rewrite theme; update cell lock behavior
lots of work on theme to use CSS instead of inline styles; more of a
hybrid now than it used to be, but a lot cleaner. the actual UX is
still up in the air.

move lock to style, so it can cascade (as noted in many places). added
lock button to toolbar by default (may need option to disable?)

update version to 9