fix scroll issue
when you focus the spreadsheet, if it has been scrolled it will pop back
to the origin because it's focusing on the overlay editor, which is offscreen.

to resolve this if you focus the spreadsheet (and the editor) and it's not
presently focused, we move it to the top-left of the scroll position (with
a small buffer).