update for mobile
we don't want to pop up the keyboard on mobile when you click the
grid. this happens because of our new overlay editor. so now, when
on mobile we focus on the container, not the overlay. that means you
can't just start typing, or use the arrow keys, but that's OK. if
we want to handle the arrow keys we can do that separately (call that
a TODO).

this requires some additional handling on double-click to start the
ICE, even though that probably won't ever happen (it would be similar
to double-clicking, probably an ipad with keyboard and mouse).