add scale controls
scale controls go in the tab bar, and if you enable scale controls tab
bar is implicit. there's a separate setting for persisting scale. we
need a key to do this, so that setting is either boolean (use default
key) or a string (use a specific key, which will be prepended with
another string).

the string key for persistence is intended to disambiguate among several
spreadsheets on the same page, which can have separate scales.

persisted scale overrides scale parameter. if you don't want this
behavior, don't persist scale.