switch to font scale on toolbar
instead of defaulting to points, we show "scale"; this is relative to
the base theme size, so most cells should show "1.00".

this solves a couple of problems with font sizing. one thing that's
still slightly broken is we're possibly setting "1em" unecssarily, that
needs to get cleaned up in the style pruning routine.

you can still type in points/px/whatever in the text box, but it will
always show scale as 2-digit fixed point.

FIXME: maybe l10n for font size? we can use parser...