localize font scale for input, output
dot to comma, basically. we use regex on parsing because it should never
be more complicated than that, and we discard values that look like
junk. there's still an edge case where you are in locale with ',' and
you type '.', we'll interpret that as decimal. this is an error but
because we do not intend to support 1,000pt font sizes, we can ignore it
for the time being.