style changes
clean up some style property keys, composite font size unit/value

we originally wanted to composite borders, but that's not useful because
we want cascading separately for border width and border color. since
the desire for the composite border objects was based on code needs, we
add a separate method for generating composite borders from a style.