restructure event source
move event source to a member, instead of a superclass (-> composition).
the aim here is to split the events, so we can remove MC events from the
regular set.

we can do that by using local members for publish/subscribe, and then
overloading the type in the subclass. there's one limitation that we
have to | the type in the subclass subscribe method, but this seems like
a reasonable limitation.

all this is because we can't extend the type.