more work on decorated functions
it's now generic, and parameterized in functions, so theoretically we
can add functions just by setting a flag.

while working on this I noticed some issues on required parameters -- if
we don't require the same parameters the resulting function will break
XLSX. we could resolve this with a similar set of flags (if no value,
populate the default, which we have).