updates to support (separate) compilation of mjs
as a proper module. the only real holdup is the inability to detect the
script path, but if you are loading as a module you can set this
explicitly via a method on the TREB object.

we also rename all the TREB scripts, and default to es2020 instead of
defaulting to es5, so the conditional loaders have to change as well.