update for annotation layout
(1) if the corner of an annotation was out of bounds, the method used to
find the layout would return row or column = 0; that broke XLSX export.
we now support overflow in that method.

(2) we also want to ensure the sheet extent is >= the corner cell. we do
that on AddAnnotation and move/resize, which covers some cases; we
probably still need to do it for some other methods, including
add/remove row/column and possibly some others.