update MC calculator to fix prep errors
if a random distribution was called in an IF function during the
simulation prep pass, and the function was not called because the branch
was skipped when evaluating the IF function, then it would not be set up
properly and subsequent operations would fail.

not sure why this does not happen in Excel -- do they always evaluate
both branches of IF functions?

also, are there any other cases where it might branch? (...) sort of
immaterial since we write the code to skip branches, and we only do it
(for now) for IF functions.