update for better caching support
some RAW projects want to cache the spreadsheet by moving elements. the
composite sheet was caching the "container" element, which makes sense,
but we were switching that in the RAW cache. so the solution is to
dynamically look up the container when toggling sidebar/toolbar.

this isn't necessarily the best solution, but it's nondestructive and
solves the problem for the time being.