mark ctor as internal
even though it's not private. there are two reasons: one we don't want
anyone calling it directly, and two we want to mask the DataModel type
which is a parameter (actually a type of a parameter of a parameter).
otherwise that gets in the generated API typings and creates a mess.