QoL: fix increase/decrease decimal from General
general has a magic decimal, and 0-6 decimal places. previously if you
increased/decreased decimal it would do it from zero. what we want (and
what this change does) is update from the actual displayed value.

this is a little complicated because we have to format it to check, but
it's a button press so we don't have to be super efficient.

we base this on the first number encountered, which might not always be
exactly what you want but it's better than the old behavior.