update to fix some MC issues
we have a cache for simulation results called "last_simulation_data".
this used to be in EmbeddedSpreadsheet. because it's MC specific, it is
defined in the MC version only.

this issue that came up is that we were setting this field on document
load, and document load was (in some cases) called from the base-class
constructor.

what happens is the field is initialized by the subclass constructor
_after_ the super() call, so the field is unset (reset) following the
initial load method.

by moving this to the calculator class, we can resolve this timing issue
(even though the calculator is also a subclass) because the calculator
is fully initialized before the EmbeddedSpreadsheet constructor is
called, and therefore before the load method.

all this makes sense in retrospect but it was confusing, and it happened
because of the MC/no-MC split.