switch to fetch and fix cache behavior
we were using XHR instead of fetch because of IE11. drop that and start
using fetch, and while we're at it, fix cache behavior.

the default in XHR was to not use cache. the default in fetch is to use
cache. we want to use it. rationale: that's what it's for, these are
just regular network resources. the server sending the documents can
manage cache how they see fit.