track file version with file
so we have a consistent record we can use for dirty/revert checks. this
is just an opaque number, starting at 0, which increments on any change
to the document -- structure or data, but not calculation cell values
(so a change to the output of RAND() or TODAY() is not considered a
document change).