working on coediting but fixed a lot of issues
along the way -- specifically in patching functions when adding/removing
rows, and handling references to deleted sheets better.

we _could_ expressly rewrite functions when we delete a sheet, which
might be useful in that they would recalculate. not sure. Excel has the
ability to insert the name #REF into dead references, which is good
because you don't accidentally re-create the reference if you add a
sheet with the (now deleted sheet) name.

but we can't name the sheet that because the parser doesn't like it.
note that Excel's parser doesn't like it either, but it allows it for
existing functions (doesn't rewrite). it's magic. we can do magic.