import/export fixes; update OFFSET function
OFFSET had arguments in the wrong order (width, height) should be
(height, width).

import/export and clean up named expressions. when importing/exporting
we need to make sure we have the sheet name (export) or the ID (import)
and for export we also want to make sure addresses are absolute.