embed import/export worker in es module
traditionally we split the import/export functions into a separate
worker script, which is loaded on demand. that's still the case for the
js lib.

to simplify consuming the es module, we're embedding the worker in the
generated .mjs script. it's loaded as a blob when required. this makes
the es module quite a bit larger, but it means the worker doesn't have
to be publicly available on the webserver.

TODO: we might build separately in the event the worker is not needed.