clean up and clarify revert pattern
network_document supports loading network documents. storage_key
supports storing local changes in localStorage. we added a new option
for documents supporting reversion, but that is not necessary -- we
should just use the original options.

1 - if you only ever want the canonical version, use network_document
without storage_key

2 - if you want the local version to override, but only if available,
use both

3 - if you want to support reversion, you don't have to do anything
(except maybe turn on the revert button).

the original behavior was to always load network_document even if
storage_key was set (and a document was in storage); but this make no
sense, because then there's no reason to ever set the two options
together. so we can use this to support reversion without requiring a
new option.

leaving the old option "alternate_document" in place for backcompat but
it's deprecated.