Coerce calculated values to 0 from undefined
we were allowing undefined values to propogate, if the result of a
calculation was a reference to an empty cell (for example if the formula
was `=A2` and A2 was empty). this should more properly resolve to 0, as
it is the result of a formula.

this only affects the result of a formula; it doesn't change how an
empty cell is treated as an argument to a cell function -- for example,
if A2 is empty then `=CONCATENATE(A2, "OK")` resolves to "OK" --
undefined is treated as an empty string in this case.