support for import/export of tables
basic, at the moment. we drop table styling on import, since we only
have a single table style.

there are (apparently) constraints on table column headers/column names.
they have to be strings, and they have to be unique
(case-insensitively). we are not enforcing that in the spreadsheet, at
the moment, so we need to support it on export or it will break.

two things to add to table TODOs:

 - enforce column naming restrictions, as above

 - add support for table references. these apparently only work in the
   same row, and they reference the table name and column name. not sure
   why these are useful, maybe something they inherited?

   ok, table references are more complicated than I thought (and maybe
   slightly more useful?) see

  https://support.microsoft.com/en-us/office/using-structured-references-with-excel-tables-f5ed2452-2337-4f71-bed3-c8ae6d2b276e