add SUBTOTAL function
this function is for tables, and uses numbers to switch between
different stats functions. the significant thing about this function is
that it ignores hidden rows. this is not standard spreadsheet behavior,
and it makes building and supporting this function more complicated.

it's also fairly expensive, so it should probably not be used unless
absolutely necessary.