clearing area removes any contained tables
if you clear cells without changing structure (i.e. select an area and
delete, but don't delete columns or rows) and the area completely
contains one or more tables, we will remove those tables. if you only
select part of a table, we remove the contents but not the table
structure.