moving to custom element
after the dev/testing work we can move to a custom element, although we
are not going to use shadow DOM for reasons outlined elsewhere. along
with that change we're going to make some initial stabs at cleaning up
the node structure, although we don't need to finish that before
releasing the custom element.