updates to custom element
hooked up toolbar, dialog, sidebar; working on cleaning up layout/node
structure and classes.

we will likely do a release before we finish cleaning up the class
structure, because it's usable now and this is a better platform for
going forward. we've already resolved a few minor issues and I'd rather
use this version than backport the fixes.