restructure files
mostly moving all stylesheets to treb-embed; even though many of the
styles apply to things in treb-grid. it's kind of an artificial
distinction anyway; they are pretty closely related. having the styles
in one place makes it easier to unify, prune, and clean up the styles
(which we are doing).