clean up annotation
as part of the effort to clean up exported types, split all persisted
data out of the annotation class and into an interface. we might in the
future drop the annotation class in favor of a type, since it doesn't
actually do anything, but at least now we can expose the serialized
type.