handle inline-document="true"
we were expecting the inline-document attribute to either have a string
value or have no value; but working on an astro project we discovered
that astro explicitly sets inline-document="true" when the source has no
attribute value.

so we handle this case explicitly. this might cause problems if you want
to set your document name to the string "true", but that's an edgier
case.