update worker build to fix tsc issues
previously we were relying on tsc not actually resolving our worker
script, but probably because of differences in the way paths are
resolved on windows, that didn't work.

I cleaned up the worker processing a little bit so that now we can
actually import the file in the main script -- which is good, as it will
typecheck in the tsc pass. the worker is still inlined in the esbuild
pass. there's one minor issue with typing that requires an any when
constructing the worker, but I'll see if I can clean that up.