update for better type export
this update fixes export for TREBDocument, which is the serialized
document type. previously we were masking this type in the API.

the biggest problem with exporting this type was internal enums. we used
enums for typing data validation and cell values.

I dropped the data validation type completely, and switched it to a
string enum. this is pretty clean although we have to check for the old
type when importing.

the cell value type is a little more complicated, because it's pretty
deeply ingrained. for the time being I'm leaving that as a numeric enum
internally, but we use a string enum for the exported type. types are
not exported by default; it's kind of a special case. if you do export
types, the string values will bloat the data a bit but should compress
well.