remove the Style namespace
the Style namespace was a relic of some old code. it's finally removed.
the important types are moved into the top-level. functions in the
namespace are moved into an object with the same name, but these
functions were all marked @internal so this should have no impact on
external users.

this may break existing code. to upgrade, add imports for the specific
types (probably just CellStyle, which replaces Style.Properties) and
then you can just replace the token.

we also change the old exported `Color`, which was a single instance of
`ColorFunctions`, renamed to `ColorFunctions` which makes more sense.