updates for external editor support
- clean up selections after removing editor
- force qualified address resolution on selection

the latter has to do with relative/qualified addressing. the cell
editors switch to relative addresses if the target is in the same sheet.
to support that we'd need to add a "base address" to the external
editor, but that should be pretty simple. for now we just default to
qualified addresses.