change priority when inferring number format
when inferring a number format, we look at cell references and function
arguments. previously we'd just take the first number format. however
when one argument is a percentage and another is not, usually you do not
want to infer the percentage. so we add priority here for non-percentage
number formats.