minor optimizations
gradient changed to use a function. this function is marked as internal
so it won't show up in tooltips (although if you use it in a sheet it
will work). add an 'updated' flag to leaf vertices, and potentially skip
some checks for conditional formats based on the flag value.