clean up & fixes
clean up - remove some unecessary save-to-local-storage calls. stop
using `beforeunload` to save at session exit -- switch to
`visibilitychange`. this should be widely supported in 2023.

fix for firefox removing spaces in the overlay editor. this was
something we'd seen before, but it got missed when we rewrote the
editor. the fix is to use `whitespace:pre` on the ICE. need to monitor
that for issues.