fix issue with inferring number format
TREB tries to infer a number format if you enter a formula in a cell
that has no format set. If the inferred format was "General", we were
setting this as a specific format (i.e. 0.#####) instead of using the
symbolic name. Because magic decimal points are not represented in the
text format, we were losing the magic decimal point.

This is fixed, but the real solution would be to find a way to represent
that magic decimal point in the text format. TODO.