split chart renderer and allow overload
our old charts took in data and painted nodes. we now split this into
separate concerns so the rendering can be overloaded. if you overload
the renderer you'll get a common data format and the target node.

originally the plan was to allow overloading the functions, but since we
handle export/import it's probably useful to maintain consistent
functions; therefore we allow overloading of the rendering part. that
reduces flexibility a little bit but probably better in the long run.