clean up array application
move from individual functions to descriptor. the library will handle
wrapping the function appropriately (and we wrote a more generic
function).

it's starting to seem like this should be the default, rather than the
exception, but we don't need to get there right away. for now set the
descriptor flag and we'll apply.