adding a note regarding spill references
in some cases it's not an error to use a spill reference to refer to a
single cell? not sure what the rules are. seems to have something to do
with function vs literal, although it's not an error even if the
function does not return an array (maybe if it returns an LPXLOPER12 it
allows it because it _could_ be an array? investigate)