move import/export worker to a separate library
using import.meta.url we can load it relative to our base library. this
reduces the main library by 210k or so, and only loads it if necessary.
as a result we can drop the old "light" version which excluded
import/export functions.

loading the module requires some hacky workarounds for esbuild and tsc,
plus one for vite for downstream bundlers.