update dynamic worker loading
in this version we create a module that has an embedded worker script,
and we create the worker from that script. the module is dynamically
loaded so we don't have to load it immediately.

creating the module script and loading that as a worker doesn't work in
at least webpack (works ok in vite though, as long as we use a dummy
import).