allow markdown in comments
even if it's disabled in cells. we also update the markdown parser to
allow two newlines in a row (but only two).

there's still an issue with our md parser in that it will insert a
newline when you type a newline, even if the prior line doesn't end with
two spaces. but we prefer it this way.