handle importing spill arrays + xlfn.single
we can handle the explicit _xlfn.SINGLE functions, but there are
apparently also implicit non-dynamic arrays; not sure how these are
meant to be represented, unless it's through the explcit absence of
cell metadata.

if it is through the absence of cell metadata, then this presumably
needs to get preserved so we know in the future, when calculating, that
this should be the case? that seems impractical.

TODO