support importing 1-cell dynamic arrays
we still have to figure out how to determine when a function should have
an @. from what I can tell xlsx does not encode this; rather, it encodes
the opposite. that means that either the function itself has a flag that
says "yes we can return an array", or they test the function to see if
returns an array or not, and then apply the @ as necessary.

the first option (indicating when a function might return an array)
means a lot of back-patching functions. the alternative means some ugly
extra calculation.